Update `updateUserSettings` function in `userController.ts` to match string `_id`s
Modified the query filter in the `db.collection('users').updateOne()` call within the `updateUserSettings` function. Changed the filter from `{ _id: new ObjectId(userId) }` to `{ _id: userId }` to ensure compatibility with `_id`s stored as strings in the database, allowing the update operation to correctly identify and modify the user document.